+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -750,7 +750,7 @@ internal constructor(
     ): Call<ConnectionData> {
         return CoroutineCall(clientScope) {
             logger.d { "[switchUser] user.id: '${user.id}'" }
-            notifications.deleteDevice() // always delete device if switching users
+            notifications.deleteDevice(getCurrentUser()) // always delete device if switching users
             disconnectUserSuspend(flushPersistence = true)
             // change userId only after disconnect,
             // otherwise the userScope won't cancel coroutines related to the previous user.
@@ -1480,7 +1480,7 @@ internal constructor(
             when (isUserSet()) {
                 true -> {
                     if (deleteDevice) {
-                        notifications.deleteDevice()
+                        notifications.deleteDevice(getCurrentUser())
                     }
                     disconnectSuspend(flushPersistence)
                     Result.Success(Unit)
@@ -1554,7 +1554,7 @@ internal constructor(
 
     @CheckResult
     public fun deleteDevice(device: Device): Call<Unit> {
-        return api.deleteDevice(device)
+        return api.deleteDevice(device.token)
             .share(userScope) { DeleteDeviceIdentifier(device) }
     }
 
@@ -3617,6 +3617,8 @@ internal constructor(
      * @param memberIds The list of the member ids to be added.
      * @param systemMessage The system message that will be shown in the channel.
      * @param hideHistory Hides the history of the channel to the added member.
+     * @param hideHistoryBefore Hides the channel history before the provided date from the added members. If
+     * [hideHistory] and [hideHistoryBefore] are both specified, [hideHistoryBefore] takes precedence.
      * @param skipPush If true, skips sending push notifications.
      *
      * @return Executable async [Call] responsible for adding the members.
@@ -3628,12 +3630,14 @@ internal constructor(
         memberIds: List<String>,
         systemMessage: Message? = null,
         hideHistory: Boolean? = null,
+        hideHistoryBefore: Date? = null,
         skipPush: Boolean? = null,
     ): Call<Channel> {
         val params = AddMembersParams(
             members = memberIds.map(::MemberData),
             systemMessage = systemMessage,
             hideHistory = hideHistory,
+            hideHistoryBefore = hideHistoryBefore,
             skipPush = skipPush,
         )
         return addMembers(channelType, channelId, params)
@@ -3660,6 +3664,7 @@ internal constructor(
             members = params.members,
             systemMessage = params.systemMessage,
             hideHistory = params.hideHistory,
+            hideHistoryBefore = params.hideHistoryBefore,
             skipPush = params.skipPush,
         )
     }
@@ -5170,7 +5175,9 @@ internal constructor(
          */
         @Throws(IllegalStateException::class)
         internal fun setDevice(device: Device) {
-            ensureClientInitialized().notifications.setDevice(device)
+            val client = ensureClientInitialized()
+            val user = client.getCurrentUser()
+            client.notifications.setDevice(user, device)
         }
 
         @Throws(IllegalStateException::class)