+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api/ChatApi.kt
@@ -126,7 +126,7 @@ internal interface ChatApi {
     fun addDevice(device: Device): Call<Unit>
 
     @CheckResult
-    fun deleteDevice(device: Device): Call<Unit>
+    fun deleteDevice(id: String): Call<Unit>
 
     @CheckResult
     fun getDevices(): Call<List<Device>>
@@ -420,6 +420,7 @@ internal interface ChatApi {
         members: List<MemberData>,
         systemMessage: Message?,
         hideHistory: Boolean?,
+        hideHistoryBefore: Date?,
         skipPush: Boolean?,
     ): Call<Channel>