+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/plugin/logic/channel/internal/ChannelStateLogic.kt
@@ -784,6 +784,8 @@ internal class ChannelStateLogic(
             }
             mutableState.recoveryNeeded = true
         }
+        mutableState.setLoadingOlderMessages(false)
+        mutableState.setLoadingNewerMessages(false)
     }
 
     /**